% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airtable.R
\name{airtable}
\alias{airtable}
\title{Get data from Airtable}
\usage{
airtable(base_name = NULL, table_name = NULL, base_id = NULL,
  tr_base = FALSE, as_tibble = TRUE)
}
\description{
Get data from Airtable
}
