% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desktime.R
\name{desktime_api_period}
\alias{desktime_api_period}
\title{Desktime api data for a given period}
\usage{
desktime_api_period(period_start, period_end = Sys.Date())
}
\description{
Desktime api data for a given period
}
