% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airtable.R
\name{airtable}
\alias{airtable}
\title{Get data from Airtable}
\usage{
airtable(base_name = NULL, table_name = NULL, base_id = NULL,
  single_link = character(0), datetime = character(0),
  boolean = character(0), date = character(0))
}
\description{
Get data from Airtable
}
